Print Fibonacci sequence

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
  */
 int main(void)
 {
-	long long int x = 1;
-	long long int y = 2;
-	long long int z;
+	long int x = 1;
+	long int y = 2;
+	long int z;
 	int count = 2;
 
 	printf("%d, %d, ", x, y);
